requirements.txt
pdf2image
pytesseract
Pillow
opencv-python